Move NDRDescT and ArgDesc to separate headers
