+++ sycl/source/detail/kernel_data.hpp
@@ -8,9 +8,10 @@
 
 #pragma once
 
-#include <detail/cg.hpp>
 #include <detail/device_kernel_info.hpp>
 #include <detail/graph/dynamic_impl.hpp>
+#include <detail/kernel_arg_desc.hpp>
+#include <detail/ndrange_desc.hpp>
 
 #include <sycl/detail/kernel_desc.hpp>
